[CE-957] add routing for addresses
